v2 фикс замечаний по ревью
