+++ src/components/app/index.tsx
@@ -1,0 +1,1 @@
+export { App } from './App'
